lib: Add lib/cluster to avoid circular imports from utils
